+++ src/main/java/bot/datasource/repositories/StatisticsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Date;
 
-public interface StatisticsRepository extends JpaRepository<Statistic, Integer> {
+public interface StatisticsRepository extends JpaRepository<Statistic, Long> {
 
     @Query("SELECT statistic FROM Statistic statistic WHERE statistic.date = ?1")
     Statistic getByDate(Date date);
