+++ src/main/java/bot/datasource/services/JpaRepositoriesService.java
@@ -53,7 +53,7 @@ public class JpaRepositoriesService implements DBService {
     // posts
 
     @Override
-    public Post getPostById(Integer id) {
+    public Post getPostById(Long id) {
         return postsRepository.findById(id).orElse(null);
     }
 
@@ -98,7 +98,6 @@ public class JpaRepositoriesService implements DBService {
                 .mapToInt(Post::getLikesCount)
                 .sum();
     }
-
     @Override
     public int getLikesTop(BotUser user) {
         List<BotUser> top = usersRepository.findAll().stream()
